apps/ai/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.0.0
transformers>=4.45.0
accelerate>=0.26.0
